requirements.txt
sys
os
re
hashlib
argparse
struct
time
platform
cffi==1.15.1
cryptography==40.0.2
pyAesCrypt==6.0.0
pycparser==2.21
pycrypto==2.6.1
